WIP: New expors API
- Renamed base classes to *Collector to make their purpose a bit clearer
- Removed all concrete collectors but added a single collector collector
- Added stub Exports class which manages the collection process
- The removed concreate collectors were added to the new collectors package,
  including a few stub ones
