+++ librarian/core/collectors/hooks.py
@@ -1,0 +1,5 @@
+from ..exports import ListCollector
+
+
+class Hooks(ListCollector):
+    pass
