+++ librarian/core/collectors/middleware.py
@@ -1,0 +1,5 @@
+from ..exports import DependencyCollector
+
+
+class Middleware(DependencyCollector):
+    pass
