+++ librarian/core/collectors/plugins.py
@@ -1,0 +1,5 @@
+from ..exports import DependencyCollector
+
+
+class Plugins(DependencyCollector):
+    pass
