+++ librarian/core/collectors/routes.py
@@ -1,0 +1,5 @@
+from ..exports import DependencyCollector
+
+
+class Routes(DependencyCollector):
+    pass
